backend/requirement.txt
numpy==1.26.4
torch==2.3.1
transformers==4.43.3
tokenizers==0.19.1
sentencepiece==0.1.99
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
requests==2.32.3
kiwipiepy
rapidfuzz
pymysql
python-dotenv